refactor: 삭제하려는 시간, 테마에 대한 예약이 있을 경우 예외 호출 - custom exception : ReferentialIntegrityException

+++ src/main/java/roomescape/reservationTime/ReservationTimeService.java
@@ -1,6 +1,8 @@
 package roomescape.reservationTime;
 
 import org.springframework.stereotype.Service;
+import roomescape.exception.ReferentialIntegrityException;
+import roomescape.reservation.ReservationRepository;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -8,9 +10,11 @@ import java.util.stream.Collectors;
 @Service
 public class ReservationTimeService {
 	private final ReservationTimeRepository reservationTimeRepository;
+	private final ReservationRepository reservationRepository;
 
-	public ReservationTimeService(ReservationTimeRepository reservationTimeRepository) {
+	public ReservationTimeService(ReservationTimeRepository reservationTimeRepository, ReservationRepository reservationRepository) {
 		this.reservationTimeRepository = reservationTimeRepository;
+		this.reservationRepository = reservationRepository;
 	}
 
 	public ReservationTimeResponse saveReservationTime(ReservationTimeRequest request) {
@@ -28,7 +32,11 @@ public class ReservationTimeService {
 				.collect(Collectors.toList());
 	}
 
-	public void deleteReservation(Long id) {
+	public void deleteReservationTime(Long id) {
+		if(reservationRepository.countByTime(id) > 0) {
+			throw new ReferentialIntegrityException("해당 시간에 대한 예약");
+		}
+
 		reservationTimeRepository.delete(id);
 	}
 }
