+++ src/main/java/roomescape/theme/ThemeService.java
@@ -1,6 +1,8 @@
 package roomescape.theme;
 
 import org.springframework.stereotype.Service;
+import roomescape.exception.ReferentialIntegrityException;
+import roomescape.reservation.ReservationRepository;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -8,9 +10,11 @@ import java.util.stream.Collectors;
 @Service
 public class ThemeService {
 	private final ThemeRepository themeRepository;
+	private final ReservationRepository reservationRepository;
 
-	public ThemeService(ThemeRepository themeRepository) {
+	public ThemeService(ThemeRepository themeRepository, ReservationRepository reservationRepository) {
 		this.themeRepository = themeRepository;
+		this.reservationRepository = reservationRepository;
 	}
 
 	public List<ThemeResponse> findThemes() {
@@ -27,7 +31,11 @@ public class ThemeService {
 		return new ThemeResponse(themeRepository.findById(id));
 	}
 
-	public void deleteThemes(long id) {
+	public void deleteTheme(long id) {
+		if(reservationRepository.countByTheme(id) > 0) {
+			throw new ReferentialIntegrityException("해당 테마에 대한 예약");
+		}
+
 		themeRepository.delete(id);
 	}
 }
